Fix bug in figure.py involving mutability of numpy arrays
Running the FigureSeparator was generating the following warning:
"UserWarning: The given NumPy array is not writeable, and PyTorch does not
 support non-writeable tensors. This means you can write to the underlying
(supposedly non-writeable) NumPy array using the tensor. You may want to copy
 the array to protect its data or make it writeable before converting it to
a tensor. This type of warning will be suppressed for the rest of this
program."
Although img.flags reported that the array was writeable, it appears numpy
will sometimes report an array as writeable when it does not own the
underlying array. This was fixed by creating a copy of the array.