Refactor journal scraper to uses classes to represent journal families
This is a major refactor. I realized the logic necessary for each
journal was too complex to reasonably store in a dictionary, so
i changed it to a class. Now the JournalFamily class is nearly
completely agnostic to the journal type. Nature and ACS families
have been implemented and are working.

Any new journal families can be implemented simply by adding a
new subclass of JournalFamily