Create stable, end-to-end test for nature journals
This test stores relevant materials so there is no need for an internet
connection. Links are mocked using the responses library. These tests
do not test for correctness, but for similarity to a reference json.
This is useful to ensure a refactor has been successful, but useless to
ensure an accuracy improvement has been successful.