Separate subfigure type classification from extract_image_objects
This commit also removes the code that creates visualizations and
saves versions of images. For that reason, I am starting a new branch
to be used for the purpose of incorporating MongoDB. Once MongoDB is
incorporated, the image visualization capabilities will be restored
and this branch will be merged into the main branch.