Move code from figure.py into FigureSeparator class
We were already leaning toward this implementation, it only made
sense to embrace it. Now the code previously outside the class is
a lot cleaner, taking advantage of class attributes. This will
allow for the addition of a push_to_mongo method in the ExsclaimTool
abstract base class. This way users can decide to use Mongo or a
plain json.