Fix bugs that arise in repeated runs
Previously FigureSeparator tried to separate all images in target
directory. Now it only checks figures in the exsclaim json. Further
the list of articles that have been gone through is now appended to
instead of overwritten. These begin to address #1; a full solution,
however should allow the exsclaim json to be expanded with non
redundant information each in successive runs.